+++ pueue/src/client/cli.rs
@@ -392,21 +392,6 @@ https://github.com/Nukesor/pueue/issues/350#issue-1359083118"
         group: Option<String>,
     },
 
-    /// Accept a list or map of JSON pueue tasks via stdin and display it just
-    /// like \"pueue status\".
-    ///
-    /// A simple example might look like this:
-    ///
-    /// pueue status --json | jq -c '.tasks' | pueue format-status",
-    #[command(after_help = "DISCLAIMER:\n\
-        This command is a temporary workaround until a proper filtering language for \"status\" has
-        been implemented. It might be removed in the future.")]
-    FormatStatus {
-        #[arg(short, long)]
-        /// Only show tasks of a specific group
-        group: Option<String>,
-    },
-
     /// Display the log output of finished tasks.
     ///
     /// Only the last few lines will be shown by default.